pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "debide"
version = "0.1.0"
description = "Terminal-native Debian-focused IDE built with Textual."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "DebIDE Contributors" }]
dependencies = [
    "textual>=0.48,<1.0",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[project.scripts]
debide = "debide.cli:main"

[tool.setuptools.packages.find]
where = ["."]
